Task 227 work inventory logic

+++ app/src/main/kotlin/com/itrocket/union/inventoryCreate/domain/entity/AccountingObjectCounter.kt
@@ -1,8 +1,8 @@
 package com.itrocket.union.inventoryCreate.domain.entity
 
 data class AccountingObjectCounter(
-    val total: Int = 0, // общее количество ОУ со статусами "найден" и "не найден"
-    val found: Int = 0, // количество ОУ со статусом "найден";
-    val notFound: Int = 0, // количество ОУ со статусом "найден";
-    val new: Int = 0, // количество ОУ со статусом "новый".
+    val total: Long = 0L, // общее количество ОУ со статусами "найден" и "не найден"
+    val found: Long = 0L, // количество ОУ со статусом "найден";
+    val notFound: Long = 0L, // количество ОУ со статусом "найден";
+    val new: Long = 0L, // количество ОУ со статусом "новый".
 )
